requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
email-validator==2.2.0
Werkzeug==3.0.3


Authlib==1.3.1
itsdangerous==2.2.0

Flask-Migrate==4.0.7
alembic==1.13.2
requests==2.32.3
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.1
email-validator==2.2.0
Werkzeug==3.0.3

Authlib==1.3.1
itsdangerous==2.2.0

Flask-Migrate==4.0.7
alembic==1.13.2
requests==2.32.3
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
email-validator==2.2.0
Werkzeug==3.0.3


Authlib==1.3.1
itsdangerous==2.2.0

Flask-Migrate==4.0.7
alembic==1.13.2
requests==2.32.3
gunicorn==20.1.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
email-validator==2.2.0
Werkzeug==3.0.3
Authlib==1.3.1
itsdangerous==2.2.0
Flask-Migrate==4.0.7
alembic==1.13.2
requests==2.32.3
gunicorn==20.1.0